commas after "message"
Sections 4.1.1 and 4.1.2:  As it appears that "and is                                  
considered not valid otherwise" refers to the dc_cert_verify_algorithm
field and not also to the algorithm field, we suggest adding commas after
"message" in these compound sentences.  If this is not correct,
does "and is considered not valid otherwise" mean that if the
conditions are not met, the types are not valid?