Suggestion to use {::boilerplate bcp14-tagged}
We would like to replace the manually entered "key words" boilerplate text with {::boilerplate bcp14-tagged}, to (1) create the proper boilerplate paragraph in the output files and (2) tag all key words in the boilerplate paragraph, and all instances of key words in the document, with <bcp14> tags in XML output.  All key words will be properly bolded in .html and .pdf output files.